Bv utils
